validator-service/requirements.txt
fastapi
httpx
email-validator
uvicorn
python-dotenv
